+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - wasm_apis
-  - `web_sys::window().unwrap()` to `web_sys::window().expect_throw("Failed to get Window object")`
-  - `req.dyn_into().unwrap()` to `req.dyn_into().expect_throw("Failed to dynamically cast JsFuture into Response")`
+  - `web_sys::window().unwrap()` to `web_sys::window().expect("Failed to get Window object")`
+  - `req.dyn_into().unwrap()` to `req.dyn_into().expect("Failed to dynamically cast JsFuture into Response")`
 
 ## [0.2.0] - 2026-02-07